+++ test/tools/index.ts
@@ -2,7 +2,4 @@ export * from './data';
 export * from './collection';
 export * from './const';
 export * from './utils';
-
-// test IP
-export const IP = '127.0.0.1:19530';
-export const ENDPOINT = `http://${IP}`;
+export * from './ip';
