+++ test/tools/ip.ts
@@ -1,0 +1,3 @@
+// test IP
+export const IP = '127.0.0.1:19530';
+export const ENDPOINT = `http://${IP}`;
